refactor analyze_image_gemini.py
