+++ tools/analyzer/prompt.py
@@ -1,0 +1,110 @@
+# ruff: noqa
+prompt = """
+# 依頼内容
+
+これは日本の政治資金収支報告書の一部です。
+この画像から読み取れる全てのテキスト情報を抽出し、正確に構造化されたJSON形式で出力してください。
+不明瞭な箇所や読み取れない箇所は無理に推測せず、その旨を示すか省略してください。
+
+# 期待するデータ形式
+
+{
+  "year": 2025,
+  "categories": [
+    {
+      "id": "xxx1",
+      "name": "個人からの寄付",
+      "parent": "xxx66",
+      "direction": "income"
+    },
+    ...
+  ],
+  "transactions": [
+    {
+      "id": "xxx1",
+      "category_id": "xxxcc",
+      "name": "飲食代",
+      "date": "R5/6/28",
+      "value": 10200
+    },
+    ...
+  ]
+}
+
+# 抽出時の注意事項
+
+## directionについて
+
+directionは "income", "expense" のどちらかを指定してください。
+
+## categoriesについて
+
+- idは任意の文字列で構いません (例: "xxx1", "xxx2"など)
+- parentは親カテゴリのidを指定します。最上位カテゴリの場合はnullを指定してください
+- nameは以下のカテゴリから選んでください。ただし、当てはまるものがなく、明らかにカテゴリとして適切な物があった場合は、その他 ({{読み取れたカテゴリ}}) として立項して構いません。
+
+## categoryについて
+
+categoryは以下の中から選んでください。
+ただし、当てはまるものがなく、明らかにカテゴリとして適切な物があった場合は、その他 ({{読み取れたカテゴリ}}) として立項して構いません。
+
+categoryの候補
+- 個人からの寄附
+- 政治団体からの寄附
+- 法人その他の団体からの寄附
+- 組織活動費
+- 選挙関係費
+- 機関紙誌の発行事業費
+- 宣伝事業費
+- 政治資金パーティ開催事業費
+- その他の事業費
+- 調査研究費
+- 寄附・交付金
+- 光熱水費
+- 事務所費
+- 借入金
+- 人件費
+- 前年からの繰越額
+- 党費・会費
+- 備品・消耗品費
+- 翌年への繰越額
+- その他の経費
+- その他 ({{読み取れたカテゴリ}})
+
+## transactionsについて
+
+- idは任意の文字列で構いません (例: "xxx1", "xxx2"など)
+- category_idは、categoriesで定義したidのいずれかを指定してください
+- dateは年月日という項目に 6 9 30 などの区分で入っています。これは令和6年9月30日を示すので、 R6.9.30 という文字列に変換してください
+- nameは具体的な項目を、画像に記載の通り入力してください
+- valueは金額を数値で入力してください
+
+# 具体例
+
+{
+  "year": 2025,
+  "categories": [
+    {
+      "id": "xxx1",
+      "name": "個人からの寄付",
+      "parent": "xxx66",
+      "direction": "income"
+    },
+    {
+      "id": "xxx66",
+      "name": "総収入",
+      "parent": null,
+      "direction": "income"
+    }
+  ],
+  "transactions": [
+    {
+      "id": "xxx1",
+      "category_id": "xxx1",
+      "name": "田中太郎",
+      "date": "R6.6.29",
+      "value": 10000
+    }
+  ]
+}
+"""
